Move fork resolution into chain controller
Moving fork resolution into the chain controller simplifies the
responsibility of the block validator; it is now only responsible for
validating blocks.

This change also supports the rewriting of the chain controller in Rust,
as it makes it easier to draw an FFI boundary between the chain
controller and block validator.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>