Fix memory leak
Using slice::from_raw_parts() does not reclaim ownership of the pointer, and so
when the slice goes out of scope, the memory is not freed. Because Rust is
picky about how you free Strings and Vecs, specifically that the capacity is
accurate, this change keeps Strings and Vecs as they are and passes around
pointers to them, rather than trying to convert them to something else. This
requires holding onto the capacity in addition to the length, but guarantees
that the capacity is correct when the struct is dropped.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>